Code cleanup
Splits up modules into separate files for increased readability
